Move multiplatform js implementation module to samples project.
No need to update plugin as check tasks already applied to
"kotlin-multiplatform-js" plugin.
